+++ packages/sdks/angular-sdk/projects/angular-sdk/src/lib/components/descope/descope.component.ts
@@ -48,6 +48,7 @@ export class DescopeComponent implements OnInit, OnChanges {
     ref: HTMLElement
   ) => boolean | Promise<boolean>;
   @Input() client: Record<string, any>;
+  @Input() nonce: string;
   @Input() form: Record<string, any>;
   @Input() logger: ILogger;
   @Input() styleId: string;
@@ -170,6 +171,10 @@ export class DescopeComponent implements OnInit, OnChanges {
 
     if (this.client) {
       this.webComponent.setAttribute('client', JSON.stringify(this.client));
+    }
+
+    if (this.nonce) {
+      this.webComponent.setAttribute('nonce', JSON.stringify(this.nonce));
     }
 
     if (this.form) {