+++ packages/sdks/nextjs-sdk/test/index.test.ts
@@ -12,6 +12,7 @@ describe('index', () => {
 		expect(clientIndex).toHaveProperty('getRefreshToken');
 		expect(clientIndex).toHaveProperty('isSessionTokenExpired');
 		expect(clientIndex).toHaveProperty('isRefreshTokenExpired');
+		expect(clientIndex).toHaveProperty('getCurrentTenant');
 		expect(clientIndex).toHaveProperty('getJwtPermissions');
 		expect(clientIndex).toHaveProperty('getJwtRoles');
 		expect(clientIndex).toHaveProperty('refresh');