+++ packages/sdks/web-js-sdk/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.26.2](https://github.com/descope/descope-js/compare/web-js-sdk-1.26.1...web-js-sdk-1.26.2) (2025-03-06)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.38.0`
 ## [1.26.1](https://github.com/descope/descope-js/compare/web-js-sdk-1.26.0...web-js-sdk-1.26.1) (2025-03-04)
 
 ## [1.26.0](https://github.com/descope/descope-js/compare/web-js-sdk-1.25.0...web-js-sdk-1.26.0) (2025-03-04)