+++ packages/widgets/applications-portal-widget/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.4](https://github.com/descope/descope-js/compare/applications-portal-widget-0.3.3...applications-portal-widget-0.3.4) (2025-03-06)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.26.2`
 ## [0.3.3](https://github.com/descope/descope-js/compare/applications-portal-widget-0.3.2...applications-portal-widget-0.3.3) (2025-03-04)
 
 ### Dependency Updates